Copy env properties from oasp4j-sample-core, so that devon-sample works OOTB
oasp4j-sample-core filters out the env/application.property file when creating its jar.
It seems that it doesn't filter when deploying to eclipse's tomcat, but as we use
oasp4j-sample-core as a dependency in devon-sample, we get the filtered jar, meaning we
have to duplicate the properties in our sample.